RCA update UserServiceTest and fix delete method of UserService

+++ src/main/java/com/iesfranciscodelosrios/service/UserService.java
@@ -27,6 +27,7 @@ public class UserService implements iServices<UserEntity> {
 
     @Override
     public UserEntity delete(UserEntity userEntity) {
+        if (userEntity == null) return null;
         userRepository.delete(userEntity);
         return userEntity;
     }
